pyproject.toml
[project]
name = "oselot-ai"
version = "1.0.0"
description = "Comprehensive autonomous AI OSINT expert"
requires-python = ">=3.11"
authors = [
    {name = "00SEC Ghana"}
]
readme = "README.md"
license = "MIT"
keywords = ["security", "ai", "OSINT", "reconnaissance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security"
]
dependencies = [
    "agno>=2.4.7",
    "anthropic>=0.72.0",
    "google-genai>=1.47.0",
    "google-generativeai>=0.8.5",
    "openai>=2.6.1",
    "mcp>=1.21.0",
    "rich>=14.2.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy",
    "tavily-python",
    "httpx"
]

[project.scripts]
oselot = "oselot_cli:main"

[tool.setuptools]
py-modules = ["oselot_cli", "agent", "prompt", "tools"]
package-dir = {"" = "src"}

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"